requirements.txt
python-amazon-paapi==5.0.2
google-generativeai==0.8.3
python-telegram-bot==21.5
APScheduler==3.10.4
requests==2.32.3
Pillow==10.4.0
python-dotenv==1.0.1
aiohttp==3.10.5
